+++ R/as_bootstrap_design.R
@@ -256,3 +256,37 @@ as_bootstrap_design.survey.design <- function(design,
 
   return(rep_design)
 }
+
+#' @export
+as_bootstrap_design.DBIsvydesign <- function(design,
+                                             type = "Rao-Wu-Yue-Beaumont",
+                                             replicates = 500,
+                                             compress = TRUE,
+                                             mse = getOption("survey.replicates.mse"),
+                                             samp_method_by_stage = NULL) {
+
+  rep_design <- NextMethod(design)
+
+  # Replace 'variables' with a database connection
+  # and make the object have the appropriate class
+  rep_design$variables <- NULL
+  if (design$db$dbtype == "ODBC") {
+    stop("'RODBC' no longer supported. Use the odbc package")
+  } else {
+    db <- DBI::dbDriver(design$db$dbtype)
+    dbconn <- DBI::dbConnect(db, design$db$dbname)
+  }
+  rep_design$db <- list(
+    dbname = design$db$dbname, tablename = design$db$tablename,
+    connection = dbconn,
+    dbtype = design$db$dbtype
+  )
+  class(rep_design) <- c(
+    "DBIrepdesign", "DBIsvydesign",
+    setdiff(class(rep_design), c("DBIrepdesign", "DBIsvydesign"))
+  )
+
+  rep_design$call <- sys.call(which = -1)
+
+  return(rep_design)
+}