+++ tests/testthat/test-database-backed-designs.R
@@ -22,6 +22,25 @@ dclus1 <- svydesign(
   id = ~ dnum, weights = ~ pw, fpc = ~ fpc
 )
 
+# Test the data frame conversion functionality ----
+
+test_that(
+  "`as_data_frame_with_weights()` works for database-backed designs", {
+
+    expect_equal(
+      object = dbclus1 |> as_data_frame_with_weights(
+        vars_to_keep = c("dnum", "snum"),
+        full_wgt_name = "FULL_SAMPLE_WGT"
+      ),
+      expected = dclus1 |> as_data_frame_with_weights(
+        vars_to_keep = c("dnum", "snum"),
+        full_wgt_name = "FULL_SAMPLE_WGT"
+      )
+    )
+
+  }
+)
+
 
 # Test the generalized bootstrap functionality ----
 test_that(