+++ NEWS.md
@@ -1,3 +1,5 @@
+# svrep (development version)
+
 # svrep 0.5.1
 
 * New Features: