+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * Better documentation for the bootstrap methods covered by `as_bootstrap_design()`.
 
+* The function `as_gen_boot_design()` now works for database-backed survey design objects (i.e., objects with the class `DBIsvydesign`).
+
 * Minor updates and bug fixes:
   * The function `as_bootstrap_design()` now throws an informative error message when you supply an invalid value for the `type` argument.