+++ NEWS.md
@@ -1,4 +1,4 @@
-# svrep (development version)
+# svrep 0.6.0
 
 * Added a function `as_random_group_jackknife_design()` to create random-group jackknife replicates.